added ability to disable auth
