+++ server/index.ts
@@ -1,12 +1,9 @@
 import express, { Request, Response, Express } from 'express';
-import morgan from 'morgan';
 import nextjs from 'next';
-import nocache from 'nocache';
-import Cookies from 'cookies';
-import requireAuth from './middleware/authentication';
-import setupCsrf from './middleware/csrf';
+import requireAuth, { setDisableAuthCookies } from './middleware/authentication';
+import setupCsrfProtection from './middleware/csrf';
 import setSecurityHeaders from './middleware/security';
-import { DISABLE_AUTH_COOKIE, ID_TOKEN_COOKIE, OPERATOR_COOKIE } from '../src/constants';
+import setupLogging from './middleware/logging';
 
 const dev = process.env.NODE_ENV !== 'production';
 const app = nextjs({ dev });
@@ -60,50 +57,13 @@ const setStaticRoutes = (server: Express): void => {
 (async (): Promise<void> => {
     try {
         await app.prepare();
-
         const server = express();
 
-        server.use(
-            morgan('short', {
-                skip: req => req.originalUrl.startsWith('/_next') || req.originalUrl.startsWith('/assets'),
-            }),
-        );
-
+        setupLogging(server);
         setStaticRoutes(server);
         setSecurityHeaders(server);
-
-        server.use(nocache());
-
-        server.use((req, res, next) => {
-            if (
-                (process.env.NODE_ENV === 'development' || process.env.ALLOW_DISABLE_AUTH === '1') &&
-                req.query.disableAuth === 'true'
-            ) {
-                const cookies = new Cookies(req, res);
-                const disableAuthCookie = cookies.get(DISABLE_AUTH_COOKIE);
-
-                if (!disableAuthCookie || disableAuthCookie === 'false') {
-                    cookies.set(DISABLE_AUTH_COOKIE, 'true');
-                    cookies.set(
-                        ID_TOKEN_COOKIE,
-                        'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJjdXN0b206bm9jIjoiQkxBQyJ9.-1CZzSU-mmUoLn_RpWvrBKOib2tu_SXE2FQ1HmNYnZk',
-                    );
-                    cookies.set(
-                        OPERATOR_COOKIE,
-                        JSON.stringify({
-                            operator: {
-                                operatorPublicName: 'Blackpool Transport',
-                            },
-                            uuid: '0d1953f5-f57a-4ae3-8522-fef7da34c567',
-                        }),
-                    );
-                }
-            }
-
-            next();
-        });
-
-        setupCsrf(server);
+        setDisableAuthCookies(server);
+        setupCsrfProtection(server);
 
         unauthenticatedGetRoutes.forEach(route => {
             server.get(route, (req: Request, res: Response) => {
