+++ server/middleware/logging.ts
@@ -1,0 +1,10 @@
+import { Express } from 'express';
+import morgan from 'morgan';
+
+export default (server: Express): void => {
+    server.use(
+        morgan('short', {
+            skip: req => req.originalUrl.startsWith('/_next') || req.originalUrl.startsWith('/assets'),
+        }),
+    );
+};
